Re-run latest git trunnel to get fix from teor for #13577

+++ src/ext/trunnel/trunnel.c
@@ -1,4 +1,4 @@
-/* trunnel.c -- copied from Trunnel v1.2
+/* trunnel.c -- copied from Trunnel v1.4-pre
  * https://gitweb.torproject.org/trunnel.git
  * You probably shouldn't edit this file.
  */
@@ -75,7 +75,8 @@ uint64_t
 trunnel_htonll(uint64_t a)
 {
 #ifdef IS_LITTLE_ENDIAN
-  return trunnel_htonl(a>>32) | (((uint64_t)trunnel_htonl(a))<<32);
+  return trunnel_htonl((uint32_t)(a>>32))
+    | (((uint64_t)trunnel_htonl((uint32_t)a))<<32);
 #else
   return a;
 #endif
